arithmetic average works

+++ module3/homework/inner_product/distance.hpp
@@ -3,5 +3,5 @@
 
 class Distance {
 public:
-    int operator()(std::vector<int>& a, std::vector<int>& b) const;
+    float operator()(std::vector<int>& a, std::vector<int>& b) const;
 };
